Do not attempt to open a directory
According to the POSIX standard, the open() function (upon which fopen is based
on) does not return an error on trying to open a directory (setting errno to
EISDIR) if the the flag is set to O_RDONLY (i.e. opening a file read-only).
Therefore, we need to add an additional check to make sure the path isn't a
directory.

Signed-off-by: ydalton <yussef.dalton@gmail.com>